requirements.txt
pandas 
numpy
scikit-learn
matplotlib
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
Flask
Flask-Cors
notebook
xgboost
mlflow==2.2.2
fastf1
streamlit
-e .